// admin/src/generated/model/graphNode.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Embed Chat API
 * OpenAPI spec version: 1.0.0
 */
import type { GraphNodeDescription } from './graphNodeDescription';
import type { GraphNodeEntityType } from './graphNodeEntityType';
import type { GraphNodeSourceDocumentId } from './graphNodeSourceDocumentId';

export interface GraphNode {
  description?: GraphNodeDescription;
  entity_type?: GraphNodeEntityType;
  name: string;
  source_document_id?: GraphNodeSourceDocumentId;
}